Move vue-vm mock of reactive package to the mocks dir
