+++ packages/reactive/tests/__mocks__/vue-vm.ts
@@ -1,0 +1,19 @@
+import VueCompositionApi from '@vue/composition-api';
+import { Vue, isVue2, createApp, defineComponent } from 'vue-demi';
+
+let VueInstance: unknown | undefined;
+
+if (!VueInstance) {
+  if (isVue2) Vue.use(VueCompositionApi);
+
+  VueInstance = createApp(
+    defineComponent({
+      name: 'App',
+      render(createElement: Function) {
+        return createElement('div', this.$slots.default);
+      }
+    })
+  );
+}
+
+export default VueInstance;
